pyproject.toml
[project]
name = "ragentools"
version = "0.1.0"
description = ""
authors = [
    {name = "JamesChao"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "langchain-google-genai (>=3.0.0,<4.0.0)",
    "google-genai (>=1.46.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "langgraph (>=1.0.2,<2.0.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "openai (>=2.6.1,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]
